Give animations a joined file, extract simple_ai for goombas and such
